Added test for entry creation in form
